Let _REENT initialize lazily under proper locking.
This removes the race condition on reading _GLOBAL_REENT->__sdidinit that could
be modified simultaneously from another thread.

We are flushing all streams on every thread exit with _cleanup_r(_REENT) because
it does not work to do so from exit.c.  To be investigated.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2058
TEST=run_post_msg_thread_chrome_browser_test (did not check that it works)

Review URL: http://codereview.chromium.org/7491020