Exclude newlib's malloc implementation
We are adding a better malloc implementation built in the native_client
tree (into libnacl).  It's simple to make the newlib build just omit its
own malloc implementation, so we're doing that rather than changing it.

Note: this will produce a library that cannot build working applications,
until the corresponding native_client tree change goes in along with a
toolchain DEPS update that brings this change in.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2271
TEST= manual build-with-newlib works

R=pasko@google.com,jvoung@google.com

Review URL: http://codereview.chromium.org/7994014