Fix newlib configury for all nacl hosts to use -DMISSING_SYSCALL_NAMES
There is an arm-*-* pattern that matches and prevents the default pattern
from matching and adding -DMISSING_SYSCALL_NAMES, whereas for i?86-*-nacl*,
x86_64-*-nacl*, or le32-*-nacl*, that default case does match.  So we need
a generic *-*-nacl* pattern to match first so that we ensure we add the
necessary -DMISSING_SYSCALL_NAMES switch for all nacl builds.

Change-Id: I0b73723bdbada47452914f2060472fbda6965d54