Avoid setjmp-dj.h include for __native_client__
The nest of conditionals in setjmp.h is already mostly irrelevant
for NaCl.  When __native_client__ is defined, we undefine the macros
set by previous conditional blocks and define them uniformly for
__native_client__ on any machine.

The wrinkle arose that if __i386__ is defined and __unix__ is not also
defined, it tries to include "setjmp-dj.h", which has not just macros
but its own typedef and declarations.  We never want to get that file
for NaCl.  While the old nacl-gcc compiler happens to avoid this because
it predefines __unix__, the next-generation compiler (so far) does not
predefine __unix__, so it was bitten.

Change-Id: I7c763c1017ca9f98bd75bc6ae2b925cefe791a9e