newlib: Don't use -fshort-enums
The upstream newlib build uses -fshort-enums for a few files.  This
is intended to reduce code size, and it's harmless because the enums
affected are not used in at any ABI boundaries, just inside private
functions.  But on ARM, using -fshort-enums marks an object with
attributes saying it's using a non-default ABI and the linker warns
about mixing objects with and without -fshort-enums.  The possible
performance/code-size benefit of -fshort-enums is negligible in the
NaCl context, and it's more important that SDK users not get these
confusing warnings at link time.  Some unrelated whitespace/filling
changes appear as fallout from re-running Automake.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3193
TEST= none

Change-Id: I844aa8a2bebb43a8660dd7e316eaeebd3c0c574d